+++ view/frontend/web/js/model/checkout/bluepayment.js
@@ -52,7 +52,6 @@ define([
         gatewaysIds: {
             blik: 509,
             blik_bnpl: 523,
-            smartney: 700,
             hub: 702,
             paypo: 705,
             card: 1500,